move static data to file
